(#2054) Move Focus Trap to Shared Location
- The modal needs the same focus trap as the header, so we're making it shared.

Touches #2054
