+++ packages/ncids-js/src/components/nci-header/utils/mega-menu/mega-menu-nav.ts
@@ -1,4 +1,4 @@
-import { FocusTrap } from '../focus-trap';
+import { FocusTrap } from '../../../../utils/focus-trap';
 import {
 	MegaMenuAdapter,
 	MegaMenuDisplayEventDetails,
