+++ packages/ncids-js/src/components/nci-header/utils/mobile-menu/mobile-menu.ts
@@ -1,4 +1,4 @@
-import { FocusTrap } from '../focus-trap';
+import { FocusTrap } from '../../../../utils/focus-trap';
 import {
 	MobileMenuAdapter,
 	MobileMenuData,
