+++ packages/ncids-js/src/utils/__test__/focus-trap/focustrap-dom.ts
