+++ packages/ncids-js/src/utils/focus-trap.ts
